+++ admin/server/middleware/request.js
@@ -114,15 +114,22 @@ module.exports = async function (req, res, next, nextNode) {
 	frontendCookie = (localization[frontendCookie] && localization[frontendCookie].value) 
 		|| localization[defaultLanguage].value;
 
+
+	/*
+	** Use system user prefer language first
+	** Terry Chan
+	** 11/06/2019
+	*/
+	const preferLanguage = req.user && req.user.language;
 	req.locales = {
 		// localization language set
 		localization,
 		// default language to pickup the data
 		defaultLanguage,
 		// adminUI current data language, serve the any requested lang first before cookie lang
-		langd: req.headers.langd || req.query.langd || req.body.langd || dataCookie,
+		langd: req.headers.langd || req.query.langd || req.body.langd || preferLanguage || dataCookie,
 		// adminUI current layout language
-		langf: frontendCookie,
+		langf: preferLanguage || frontendCookie,
 	};
 	// }
 
@@ -132,6 +139,15 @@ module.exports = async function (req, res, next, nextNode) {
 	nextNode.set('langf', req.locales.langf);
 	nextNode.set('langd', req.locales.langd);
 	nextNode.set('language pack', req.locales.localization);
+
+	/*
+	** update the user prefer language if it is being changed
+	** 11/06/2019
+	*/
+	if (preferLanguage !== req.locales.langf) {
+		req.user.set('language', req.locales.langf);
+		req.user.save();
+	}
 	// console.log(req.menu, req.appLanguage);
 	if (next) next();
 };